# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart-ChatBox API",
    "description": "API para gerenciamento de WhatsApp, empresas, projetos e leads.",
    "version": "1.0.0",
    "contact": {
      "name": "Tiago Marins",
      "url": "https://github.com/tiagomarins-dev/smart-chatbot-mb"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor de produção"
    },
    {
      "url": "http://localhost/api",
      "description": "Servidor local"
    }
  ],
...
  "paths": {
...
    "/v1/projects": {
      "get": {
        "tags": ["projects"],
        "summary": "Listar projetos",
        "description": "Retorna a lista de projetos do usuário autenticado. É necessário fornecer o ID da empresa para filtrar os projetos.",
        "operationId": "getProjects",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "query",
            "description": "ID da empresa para filtrar projetos (obrigatório)",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "query",
            "description": "ID do projeto (opcional)",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de projetos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["projects"],
        "summary": "Criar projeto",
        "description": "Cria um novo projeto vinculado a uma empresa",
        "operationId": "createProject",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "Dados do projeto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Projeto criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/Project"
                    },
                    "message": {
                      "type": "string",
                      "example": "Projeto criado com sucesso"
                    },
                    "details": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "Campanha de Verão"
                        },
                        "company": {
                          "type": "string",
                          "example": "Milla Borges"
                        },
                        "created_at": {
                          "type": "string",
                          "example": "04/05/2025 14:30:00"
                        },
                        "campaign": {
                          "type": "object",
                          "properties": {
                            "start": {
                              "type": "string",
                              "example": "01/06/2025"
                            },
                            "end": {
                              "type": "string",
                              "example": "31/08/2025"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Empresa não encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/projects/{id}": {
      "put": {
        "tags": ["projects"],
        "summary": "Atualizar projeto",
        "description": "Atualiza um projeto existente",
        "operationId": "updateProject",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do projeto",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Dados do projeto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdateInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Projeto atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "project": {
                      "$ref": "#/components/schemas/Project"
                    },
                    "message": {
                      "type": "string",
                      "example": "Projeto atualizado com sucesso"
                    },
                    "details": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string",
                          "example": "Campanha de Verão 2025"
                        },
                        "company": {
                          "type": "string",
                          "example": "Milla Borges"
                        },
                        "updated_at": {
                          "type": "string",
                          "example": "04/05/2025 15:45:22"
                        },
                        "status": {
                          "type": "string",
                          "example": "Ativo"
                        },
                        "campaign": {
                          "type": "object",
                          "properties": {
                            "start": {
                              "type": "string",
                              "example": "01/06/2025"
                            },
                            "end": {
                              "type": "string",
                              "example": "15/09/2025"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["projects"],
        "summary": "Desativar projeto",
        "description": "Desativa um projeto existente (soft delete)",
        "operationId": "deactivateProject",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID do projeto",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto desativado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Projeto desativado com sucesso"
                    },
                    "details": {
                      "type": "object",
                      "properties": {
                        "project_id": {
                          "type": "string",
                          "format": "uuid",
                          "example": "123e4567-e89b-12d3-a456-426614174000"
                        },
                        "project_name": {
                          "type": "string",
                          "example": "Campanha de Verão"
                        },
                        "company": {
                          "type": "string",
                          "example": "Milla Borges"
                        },
                        "deactivated_at": {
                          "type": "string",
                          "example": "04/05/2025 16:30:00"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Projeto não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido através do endpoint de autenticação"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Chave API obtida na página de integrações"
      }
    },
    "schemas": {
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "user_id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "name": {
            "type": "string",
            "example": "Campanha de Verão"
          },
          "description": {
            "type": "string",
            "example": "Campanha de marketing para o verão de 2025"
          },
          "campaign_start_date": {
            "type": "string",
            "format": "date",
            "example": "2025-06-01"
          },
          "campaign_end_date": {
            "type": "string",
            "format": "date",
            "example": "2025-08-31"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-04T12:00:00Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-04T12:00:00Z"
          }
        }
      },
      "ProjectInput": {
        "type": "object",
        "required": ["name", "company_id"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Campanha de Verão"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "description": {
            "type": "string",
            "example": "Campanha de marketing para o verão de 2025"
          },
          "campaign_start_date": {
            "type": "string",
            "format": "date",
            "example": "2025-06-01"
          },
          "campaign_end_date": {
            "type": "string",
            "format": "date",
            "example": "2025-08-31"
          }
        }
      },
      "ProjectUpdateInput": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Campanha de Verão 2025"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "description": {
            "type": "string",
            "example": "Campanha de marketing atualizada para o verão de 2025"
          },
          "campaign_start_date": {
            "type": "string",
            "format": "date",
            "example": "2025-06-01"
          },
          "campaign_end_date": {
            "type": "string",
            "format": "date",
            "example": "2025-09-15"
          },
          "is_active": {
            "type": "boolean",
            "example": true
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Mensagem de erro"
          }
        }
      },
...
    }
  }
}